+++ ui/packages/shared/components/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.374](https://github.com/parca-dev/parca/compare/@parca/components@0.16.373...@parca/components@0.16.374) (2025-09-26)
+
+**Note:** Version bump only for package @parca/components
+
+## 0.16.373 (2025-09-25)
+
+**Note:** Version bump only for package @parca/components
+
 ## [0.16.372](https://github.com/parca-dev/parca/compare/@parca/components@0.16.371...@parca/components@0.16.372) (2025-09-16)
 
 **Note:** Version bump only for package @parca/components