+++ ui/packages/shared/test-utils/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.15 (2025-09-26)
+
+**Note:** Version bump only for package @parca/test-utils
+
 ## [0.0.14](https://github.com/parca-dev/parca/compare/@parca/test-utils@0.0.13...@parca/test-utils@0.0.14) (2025-09-16)
 
 **Note:** Version bump only for package @parca/test-utils